backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.30
langchain-openai==0.1.7
langchain-core==0.2.10
langchain-text-splitters==0.2.1
faiss-cpu==1.8.0
numpy==1.26.4
PyMuPDF==1.24.5
pdfplumber==0.11.0
sentence-transformers==2.7.0
tiktoken==0.7.0
python-multipart==0.0.9
pytest==8.2.1
httpx==0.27.0
